% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plp_wrapper.R
\name{run_snp_pileup}
\alias{run_snp_pileup}
\title{SNP Pileup Analysis}
\usage{
run_snp_pileup(
  vcffile,
  output,
  bamfiles,
  count_orphans = FALSE,
  gzipped = FALSE,
  ignore_overlaps = FALSE,
  min_base_quality = 0,
  min_map_quality = 0,
  min_read_counts = 0,
  max_depth = 4000,
  progress = FALSE,
  psuedo_snps = 0,
  verbose = FALSE,
  debug_mode = FALSE
)
}
\arguments{
\item{output}{Output file path.}

\item{bamfiles}{Vector of bamfile paths.}

\item{file}{Input file path (e.g., VCF file).}
}
\value{
None. Results are written to the output file.
}
\description{
Runs the SNP pileup analysis with the given arguments.
}
